Force usage of TLSv1.2 protocol for HTTPS connections
TLS protocol v1.2 is supported on Java 7, but is disabled by default, and it's
used on Java 8 and later versions by default. So, if a server only supports
TLSv1.2, then Java 7 clients would not be able to connect.

This commit ensures that TLSv1.2 and only TLSv1.2 is used to connect to figo
API endpoints both on Java 7 and Java 8. This also means that if the server
doesn't support TLSv1.2, the connection will now fail, instead of being
downgraded to earlier TLS versions.

Signed-off-by: Yury V. Zaytsev <yury.zaytsev@moneymeets.com>